camera scroll over multiple area with proper bound checking

+++ utils.py
@@ -12,16 +12,16 @@ def screenToWorld(position, cameraPos):
     x, y = position
     lx, ly = x - config.origin[0] + cameraPos[0], y - config.origin[1] + cameraPos[1]
     cx, cy = lx // config.tileSize[0], ly // config.tileSize[1]
-    tx, ty = lx % config.tileSize[0], ly % config.tileSize[1]    
+    tx, ty = int(lx % config.tileSize[0]), int(ly % config.tileSize[1])
     return (cx + cy, cy - cx), tx, ty
 
 def localToIndex(position):
     x, y = position
-    return x + y * config.mapSize[0]
+    return int(x + y * config.mapSize[0])
 
 def worldToArea(position):
     x, y = position
-    return x // config.mapSize[0], y // config.mapSize[1]
+    return int(x // config.mapSize[0]), int(y // config.mapSize[1])
 
 def worldToLocal(area, worldPos):
     x, y = worldPos
